+++ src/main/java/kitchenpos/products/domain/Product.java
@@ -1,6 +1,9 @@
 package kitchenpos.products.domain;
 
+import kitchenpos.products.infra.PurgomalumClient;
+
 import javax.persistence.Column;
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
@@ -14,13 +17,19 @@ public class Product {
     @Id
     private UUID id;
 
-    @Column(name = "name", nullable = false)
-    private String name;
+    @Embedded
+    private DisplayedName name;
 
-    @Column(name = "price", nullable = false)
-    private BigDecimal price;
+    @Embedded
+    private Price price;
 
     public Product() {
+    }
+
+    public Product(final String name, PurgomalumClient purgomalumClient, final BigDecimal price) {
+        this.id = UUID.randomUUID();
+        this.name = new DisplayedName(name, purgomalumClient);
+        this.price = new Price(price);
     }
 
     public UUID getId() {
@@ -31,19 +40,19 @@ public class Product {
         this.id = id;
     }
 
-    public String getName() {
+    public DisplayedName getName() {
         return name;
     }
 
-    public void setName(final String name) {
-        this.name = name;
+    public void setName(final String name, PurgomalumClient purgomalumClient) {
+        this.name = new DisplayedName(name, purgomalumClient);
     }
 
-    public BigDecimal getPrice() {
+    public Price getPrice() {
         return price;
     }
 
-    public void setPrice(final BigDecimal price) {
-        this.price = price;
+    public void changePrice(final BigDecimal price) {
+        this.price = new Price(price);
     }
 }
