--- src/main/java/kitchenpos/products/tobe/application/ProductService.java
@@ -1,69 +1,0 @@
-package kitchenpos.products.tobe.application;
-
-import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.tobe.domain.ProductRepository;
-import kitchenpos.products.tobe.infra.PurgomalumClient;
-import kitchenpos.products.tobe.ui.dto.ProductCreateRequest;
-import kitchenpos.products.tobe.ui.dto.ProductCreateResponse;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-public class ProductService {
-    private final ProductRepository productRepository;
-    private final MenuRepository menuRepository;
-    private final PurgomalumClient purgomalumClient;
-
-    public ProductService(
-            final ProductRepository productRepository,
-            final MenuRepository menuRepository,
-            final PurgomalumClient purgomalumClient
-    ) {
-        this.productRepository = productRepository;
-        this.menuRepository = menuRepository;
-        this.purgomalumClient = purgomalumClient;
-    }
-
-    @Transactional
-    public ProductCreateResponse create(final ProductCreateRequest request) {
-        final Product product = new Product(request.getName(), purgomalumClient, request.getPrice());
-        Product saveProduct = productRepository.save(product);
-        return ProductCreateResponse.of(
-                saveProduct.getId(),
-                saveProduct.getName().getName(),
-                saveProduct.getPrice().getPrice())
-                ;
-    }
-//
-//    @Transactional
-//    public Product changePrice(final UUID productId, final Product request) {
-//        final BigDecimal price = request.getPrice();
-//        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
-//            throw new IllegalArgumentException();
-//        }
-//        final Product product = productRepository.findById(productId)
-//                .orElseThrow(NoSuchElementException::new);
-//        product.setPrice(price);
-//        final List<Menu> menus = menuRepository.findAllByProductId(productId);
-//        for (final Menu menu : menus) {
-//            BigDecimal sum = BigDecimal.ZERO;
-//            for (final MenuProduct menuProduct : menu.getMenuProducts()) {
-//                sum = sum.add(
-//                        menuProduct.getProduct()
-//                                .getPrice()
-//                                .multiply(BigDecimal.valueOf(menuProduct.getQuantity()))
-//                );
-//            }
-//            if (menu.getPrice().compareTo(sum) > 0) {
-//                menu.setDisplayed(false);
-//            }
-//        }
-//        return product;
-//    }
-//
-//    @Transactional(readOnly = true)
-//    public List<Product> findAll() {
-//        return productRepository.findAll();
-//    }
-}
