--- src/main/java/kitchenpos/products/tobe/domain/JpaProductRepository.java
@@ -1,8 +1,0 @@
-package kitchenpos.products.tobe.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, UUID> {
-}
