--- src/main/java/kitchenpos/products/tobe/infra/PurgomalumClient.java
@@ -1,5 +1,0 @@
-package kitchenpos.products.tobe.infra;
-
-public interface PurgomalumClient {
-    boolean containsProfanity(String text);
-}
