--- src/test/java/kitchenpos/products/tobe/application/FakePurgomalumClient.java
@@ -1,20 +1,0 @@
-package kitchenpos.products.tobe.application;
-
-import kitchenpos.products.tobe.infra.PurgomalumClient;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class FakePurgomalumClient implements PurgomalumClient {
-    private static final List<String> profanities;
-
-    static {
-        profanities = Arrays.asList("비속어", "욕설");
-    }
-
-    @Override
-    public boolean containsProfanity(final String text) {
-        return profanities.stream()
-                .anyMatch(profanity -> text.contains(profanity));
-    }
-}
