--- src/test/java/kitchenpos/products/tobe/application/ProductServiceTest.java
@@ -1,88 +1,0 @@
-package kitchenpos.products.tobe.application;
-
-import kitchenpos.menus.application.InMemoryMenuRepository;
-import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.tobe.domain.ProductRepository;
-import kitchenpos.products.tobe.infra.PurgomalumClient;
-import kitchenpos.products.tobe.ui.dto.ProductCreateRequest;
-import kitchenpos.products.tobe.ui.dto.ProductCreateResponse;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullSource;
-import org.junit.jupiter.params.provider.ValueSource;
-
-import java.math.BigDecimal;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
-
-class ProductServiceTest {
-    private ProductRepository productRepository;
-    private MenuRepository menuRepository;
-    private PurgomalumClient purgomalumClient;
-    private ProductService productService;
-
-    @BeforeEach
-    void setUp() {
-        productRepository = new InMemoryProductRepository();
-        menuRepository = new InMemoryMenuRepository();
-        purgomalumClient = new FakePurgomalumClient();
-        productService = new ProductService(productRepository, menuRepository, purgomalumClient);
-    }
-
-    @DisplayName("상품을 등록할 수 있다.")
-    @Test
-    void create() {
-        final ProductCreateRequest expected = createProductRequest("후라이드", 16_000L);
-        final ProductCreateResponse actual = productService.create(expected);
-        assertThat(actual).isNotNull();
-        assertAll(
-                () -> assertThat(actual.getId()).isNotNull(),
-                () -> assertThat(actual.getName()).isEqualTo(expected.getName()),
-                () -> assertThat(actual.getPrice()).isEqualTo(expected.getPrice())
-        );
-    }
-
-    @DisplayName("상품의 가격이 올바르지 않으면 등록할 수 없다.")
-    @ValueSource(strings = "-1000")
-    @NullSource
-    @ParameterizedTest
-    void create(final BigDecimal price) {
-        final ProductCreateRequest expected = createProductRequest("후라이드", price);
-        assertThatThrownBy(() -> productService.create(expected))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("상품의 이름이 올바르지 않으면 등록할 수 없다.")
-    @ValueSource(strings = {"비속어", "욕설이 포함된 이름"})
-    @NullSource
-    @ParameterizedTest
-    void create(final String name) {
-        final ProductCreateRequest expected = createProductRequest(name, 16_000L);
-        assertThatThrownBy(() -> productService.create(expected))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    private ProductCreateRequest createProductRequest(final String name, final long price) {
-        return createProductRequest(name, BigDecimal.valueOf(price));
-    }
-
-    private ProductCreateRequest createProductRequest(final String name, final BigDecimal price) {
-        final ProductCreateRequest request = new ProductCreateRequest(name, price);
-        return request;
-    }
-
-    private Product changePriceRequest(final long price) {
-        return changePriceRequest(BigDecimal.valueOf(price));
-    }
-
-    private Product changePriceRequest(final BigDecimal price) {
-        final Product product = new Product();
-        product.setPrice(price);
-        return product;
-    }
-}
